requirements.txt
openai==1.97.1
python-dotenv==1.0.1
pydantic==2.10.5
tqdm==4.67.1
mlflow==2.22.0
opencv-python==4.10.0.84
numpy==1.26.3
PyYAML==6.0.2
transformers==4.46.0
torch==2.5.1
Pillow==10.2.0
requests==2.32.2
moviepy==2.1.2
openai-whisper==20250625
pyrate-limiter==3.7.0
mistralai==1.6.0
